Update Docusaurus configuration for new domain and path
The configuration file for Docusaurus has been updated to reflect the new production URL and base URL for the website. This change was necessary due to the migration of the project to the 'crossfichain.github.io' domain. The 'organizationName' has been updated to 'crossfichain' to match this change. The base URL has been changed to '/docs/' to serve the site under the correct path.